pyproject.toml
[project]
name = "reid-task"
version = "0.1.0"
description = "ReID diploma repository"
authors = [
    {name = "Anna",email = "maksimova.am@phystech.edu"}
]
readme = "README.md"
requires-python = "~3.10"
dependencies = [
    "torch (>=2.7.0,<3.0.0)",
    "torchvision (>=0.22.0,<0.23.0)",
    "torchaudio (>=2.7.0,<3.0.0)",
    "yacs (>=0.1.8,<0.2.0)",
    "timm (>=1.0.15,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "fire (>=0.7.0,<0.8.0)",
    "ftfy (>=6.3.1,<7.0.0)",
    "regex (>=2024.11.6,<2025.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "pytorch-lightning (>=2.5.1.post0,<3.0.0)",
    "tensorboard (>=2.19.0,<3.0.0)",
    "tensorboardx (>=2.6.2.2,<3.0.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
